Fix for recent addition of overflow check

+++ include/CLI/TypeTools.hpp
@@ -81,7 +81,7 @@ bool lexical_cast(std::string input, T &output) {
         size_t n = 0;
         long long output_ll = std::stoll(input, &n, 0);
         output = static_cast<T>(output_ll);
-        return n == input.size() && output == output_ll;
+        return n == input.size() && static_cast<long long>(output) == output_ll;
     } catch(const std::invalid_argument &) {
         return false;
     } catch(const std::out_of_range &) {
@@ -95,12 +95,12 @@ template <typename T,
 bool lexical_cast(std::string input, T &output) {
     if(!input.empty() && input.front() == '-')
         return false; // std::stoull happily converts negative values to junk without any errors.
-    
+
     try {
         size_t n = 0;
         unsigned long long output_ll = std::stoull(input, &n, 0);
         output = static_cast<T>(output_ll);
-        return n == input.size() && output == output_ll;
+        return n == input.size() && static_cast<unsigned long long>(output) == output_ll;
     } catch(const std::invalid_argument &) {
         return false;
     } catch(const std::out_of_range &) {
